packages/qt-dax/pyproject.toml
[project]
name = "qt-dax"
version = "0.1.0"
description = "QueryTorque DAX - Power BI model analysis and DAX optimization product"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "QueryTorque", email = "support@querytorque.com"}
]
keywords = ["dax", "power-bi", "vpax", "optimization", "analysis", "querytorque"]

dependencies = [
    # Core
    "qt-shared>=0.1.0",
    "dspy-ai>=2.5.0",
    # API
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    # Templates
    "jinja2>=3.1.0",
    # CLI
    "click>=8.1.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
pbi-desktop = [
    # Windows-only Power BI Desktop connection
    "pyadomd>=0.2.0; sys_platform == 'win32'",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.27.0",
]

[project.scripts]
qt-dax = "qt_dax.cli.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["qt_dax"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
